requirements.txt
wagtail

# developer imports
django-extensions
rich
python-dotenv
invoke

# test imports
dj-database-url==0.5.0
freezegun==0.3.15
responses>=0.16.0,<0.17.0